# openapi.yaml
...
openapi: 3.0.1
...
info:
  title: pductl
  description: |
    A little Go tool to control and monitor Baytech PDUs via the serial console port.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: /v1
paths:
...
  /temperature:
    get:
      summary: Get temperature of PDU
      operationId: temperature
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                required: [temperature]
                properties:
                  temperature:
                    type: number
                    description: "Temperature [C]"
        401:
          $ref: '#/components/responses/Error'
        403:
          $ref: '#/components/responses/Error'
        500:
          $ref: '#/components/responses/Error'
...
components:
  responses:
...
    Error:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string
                description: An error message
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY